Finished tests and scripts

+++ DatabaseScript/2021-12-15_save_script_10-create_trooplevel_table.sql
@@ -7,7 +7,7 @@ CREATE TABLE `trooplevel` (
   `Attack` double DEFAULT NULL,
   `Defence` double DEFAULT NULL,
   `CarryCap` int DEFAULT NULL,
-  `Consumption` int DEFAULT NULL,
+  `Consumption` double DEFAULT NULL,
   `Speed` double DEFAULT NULL,
   `SpecialSkills` int DEFAULT NULL,
   `HP` int DEFAULT NULL,
